Start of rewrite.
