+++ plugin.yml
@@ -1,7 +1,7 @@
 name: Vehicles
 author: JaxkDev
 version: 0.1.0
-api: [3.10.0]
+api: [3.11.0]
 mcpe-protocol: [389]
 main: JaxkDev\Vehicles\Main
 website: "https://github.com/JaxkDev/Vehicles"
