+++ resources/Designs/Design_Manifest.json
@@ -1,6 +1,6 @@
 [
   {
     "name": "BasicCar",
-    "designId": "73192065-b069-4e15-897a-b6af73dbb5bd"
+    "uuid": "73192065-b069-4e15-897a-b6af73dbb5bd"
   }
 ]
