--- resources/config.yml
@@ -1,4 +1,0 @@
-version: 0
-
-vehicles:
-  show-nametags: true
