+++ src/JaxkDev/Vehicles/Exceptions/DesignException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace JaxkDev\Vehicles\Exceptions;
+
+use pocketmine\plugin\PluginException;
+
+class DesignException extends PluginException{};
