+++ src/JaxkDev/Vehicles/Main.php
@@ -14,14 +14,13 @@ declare(strict_types=1);
 
 namespace JaxkDev\Vehicles;
 
-use pocketmine\utils\Config;
+use JaxkDev\Vehicles\Exceptions\DesignException;
 use pocketmine\command\Command;
 use pocketmine\plugin\PluginBase;
 use pocketmine\command\CommandSender;
 use pocketmine\utils\TextFormat as C;
 
-use JaxkDev\Vehicles\Vehicle\Vehicle;
-use JaxkDev\Vehicles\Vehicle\VehicleFactory;
+use JaxkDev\Vehicles\Factory;
 
 class Main extends PluginBase
 {
@@ -38,20 +37,11 @@ class Main extends PluginBase
 	/** @var EventHandler */
 	private $eventHandler;
 
-	/** @var VehicleFactory */
-	public $vehicleFactory;
+	/** @var Factory */
+	public $factory;
 
-	/** @var DesignFactory */
-	public $designFactory;
-
-	/** @var String|String[]|String[] */
+	/** @var String|String[] */
 	public $interactCommands = [];
-
-	/** @var Config */
-	private $cfgObject;
-
-	/** @var array */
-	public $cfg;
 
 	public function onLoad()
 	{
@@ -63,16 +53,16 @@ class Main extends PluginBase
 
 		//Add handlers and others here.
 		$this->commandHandler = new CommandHandler($this);
-		$this->vehicleFactory = new VehicleFactory($this);
-		$this->designFactory = new DesignFactory($this);
+		$this->factory = new Factory($this);
 		$this->eventHandler = new EventHandler($this);
 
 		//Load any that need to be loaded.
-		$this->designFactory->loadAll();
-
-		$this->cfgObject = $this->getConfig();
-		$this->cfg = $this->cfgObject->getAll();
-		$this->getLogger()->debug("Loaded Config file, Version: {$this->cfg["version"]}");
+		try{
+			$this->factory->loadDesigns();
+		} catch (DesignException $e){
+			$this->getLogger()->critical($e->getMessage());
+			$this->getServer()->getPluginManager()->disablePlugin($this);
+		}
 
 		$this->getLogger()->debug("Loaded all resources !");
 	}
