+++ src/JaxkDev/Vehicles/Vehicle/BasicCar.php
@@ -3,10 +3,10 @@
  * Vehicles, PocketMine-MP Plugin.
  *
  * Licensed under the Open Software License version 3.0 (OSL-3.0)
- * Copyright (C) 2019 JaxkDev
+ * Copyright (C) 2019-2020 JaxkDev
  *
  * Twitter :: @JaxkDev
- * Discord :: Jackthehaxk21#8860
+ * Discord :: JaxkDev#8860
  * Email   :: JaxkDev@gmail.com
  */
 
@@ -23,6 +23,9 @@ use pocketmine\nbt\tag\CompoundTag;
 use pocketmine\network\mcpe\protocol\types\SkinData;
 
 class BasicCar extends Vehicle {
+
+	//TODO Delete. (here for ref of values)
+
 	public $width = 3; //rough, probably no where near.
 	public $height = 2;
